event: deserialise strings as UTF-8
Ensure that strings being deserialised from a serialised Event are _always_
decoded as UTF-8, instead of being effectively decoded as BINARY (which in
Ruby is aliased to ASCII).

WARNING: although our codecs attempt to ensure that all `RubyString` are
encoded with UTF-8, there is presently no safeguard to ensure that the literal
bytes in the `RubyString` are valid UTF-8 before being serialised. If plugins
produce `RubyString` that use an alternate encoding scheme, and put those into
an `Event`, they will not survive a serialisation round-trip.

Fixes: https://github.com/elastic/logstash/issues/9167