substitution variables: environment before secret store
The secret store is known to have by-design performance issues, adding ~2s of
overhead to each invocation; in pipelines that extensively use substitution
variables, this can translate to minutes or hours of overhead to each pipeline
load or reload -- _even if_ they provide said variables in their environment
and avoid using the secret store altogether.

This commit re-orders the search path we take to prefer the environment,
providing a way to short-circuit this overhead in the short term.