flow metrics: provide primitive-optimized capture implementations
As we add more metrics to cover more things, and extended metrics to track
moving averages each of which need potentially hundreds of data-points, we
are entering a state where the scale of memory used to present these metrics
is non-negligible. Although a normal process may only have a few dozens of
flow metrics now (and therefore a few thousands of captures), a future
introduction of just two or three plugin-level metrics on a process with a
few thousand plugins will put us in the territory of several million
captures alive at any given time.

          unoptimized: 16+(8+24+24) = 72b
  partially optimized: 16+(8+ 8+24) = 64b
      fully optimized: 16+(8+ 8+ 8) = 40b