plugins: include ALL target platforms in lockfile
By including ALL target platforms, we can ensure that the dependency graph
includes the combined resolution of ALL of those platforms, which ensures
that bundler does not need to re-resolve dependencies when being run on ANY
of the included platforms. This eliminates the risk that there will be a
divergence between the gems listed in the checked in lock-file and those
that are resolved for packaging.